Update gls.py
After some digging in the code, I realized that this exception is probably misplaced and this fix enables to use fast method properly. The old placement of the exception forced the user to use `freq` parameter which will cause invalid results, since the fast method relies on the default frequency spacing.